raw_bank_customers data
This data contains customer information of a bank, all packed up in one cell of an excel file, with up to 21 attributes and 41,128 samples. Sample attributes are:
-  Age
- Job
- Account balance
- Number of contacts performed before this campaign and for this client 
- Outcome of the previous marketing campaign
Etc… 

Objective
Clean the data and structure it into a complete csv, which can directly be used as input data, ready for training.